Safari-609.1.11.1/LayoutTests/css3/css-supports-document-parser-context.html

<!DOCTYPE html>
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body>
<script>
description("Makes sure that CSS.supports() is parsed within the context of the owning document.")

document.test1 = undefined;
document.test2 = undefined;

window.jsTestIsAsync = true;
</script>
<iframe id="frame1" src="resources/css-supports-document-parser-context-strict.html"></iframe>
<iframe id="frame2" src="resources/css-supports-document-parser-context-nonstrict.html"></iframe>
<script>
var frame1 = document.getElementById("frame1");
var frame2 = document.getElementById("frame2");

var counter = 0;
function listener() {
	++counter;
	if (counter == 2) {
		shouldBeFalse("document.test1");
		shouldBeTrue("document.test2");
		finishJSTest();
	}
}
frame1.addEventListener("load", listener);
frame2.addEventListener("load", listener);
</script>
<script src="../resources/js-test-post.js"></script>
</body>
</html>
