LayoutTests/css3/css-supports-document-parser-context-expected.txt

Makes sure that CSS.supports() is parsed within the context of the owning document.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.test1 is false
PASS document.test2 is true
PASS successfullyParsed is true

TEST COMPLETE
